linking js files and adding register page

+++ public/assets/js/about.js
@@ -1,6 +1,13 @@
 $(function(){
-	$("#navbarDropdown").text("Welcome, "+sessionStorage.userName);
-	$("#logOut").on("click", function(){
-		sessionStorage.clear();
-	})
+	if(sessionStorage.userName){
+		$("#navbarDropdown").text("Welcome, "+sessionStorage.userName);
+		$("#logOut").on("click", function(){
+			sessionStorage.clear();
+		});
+	}
+	else{
+		$("#navbarDropdown").text("Welcome, Guest");
+		$(".dropdown-item").text("Register");
+		$(".dropdown-item").attr("href", "/register");
+	};
 })
